pyproject.toml
[project]
name = "llm-engineer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.3.7",
    "datasets==3.6.0",
    "feedparser>=6.0.12",
    "google-generativeai>=0.8.6",
    "gradio>=6.2.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "jupyter-dash>=0.4.2",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.4",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "matplotlib>=3.10.8",
    "modal>=1.3.0.post1",
    "numpy>=2.4.0",
    "ollama>=0.6.1",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "protobuf==3.20.2",
    "psutil>=7.1.3",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "sentence-transformers>=5.2.0",
    "setuptools>=80.9.0",
    "speedtest-cli>=2.1.3",
    "tiktoken>=0.12.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "wandb>=0.23.1",
]
